chore: migrate monetary fields from Double to BigDecimal and enhance account management
Refactored the core data models and UI components to use `java.math.BigDecimal` instead of `Double` for all currency and amount-related fields. This ensures financial precision and avoids rounding errors across the app. Additionally, improved account management with a new "Include in Totals" feature.

- **Data Precision Migration:**
    - Updated Room entities (`AccountEntity`, `TransactionEntity`, `GoalEntity`, `SubscriptionEntity`, `CategoryEntity`, etc.) to store amounts as `BigDecimal` using new type converters.
    - Updated domain models and repositories to use `BigDecimal` for balances, transaction amounts, goals, and budget limits.
    - Adjusted `MockDataSeeder` and `MockCurrencyConverter` to handle `BigDecimal` calculations and mock data generation.
    - Updated Room database schema to version 6, reflecting the change from `REAL` to `TEXT` (stored as String for `BigDecimal`) in the database.

- **Account Management Enhancements:**
    - Added `excludeFromTotal` support in the `AddAccount` flow, allowing users to toggle whether an account's balance contributes to the overall net worth.
    - Introduced a "Include in Totals" switch in the `AddNewAccountScreen`.
    - Updated `BalanceAmount` and `AccountRow` components to visually distinguish accounts excluded from totals (using muted text colors).

- **UI & Component Updates:**
    - Integrated `PremiumCard` and `MizanIcon` into `PremiumTotalBalanceCard` for a more consistent design system.
    - Introduced `PremiumToggle` and updated `CurrencyScrollSelector` styling.
    - Refactored various dashboard and summary widgets (`BalanceCard`, `StatsRow`, `MonthlySummaryCard`) to process `BigDecimal` values for display.
    - Cleaned up unused components by deleting `AccountCard.kt` and `AddEditAccountSheet.kt`.

- **State Management:**
    - Updated MVI stores (`AccountsStore`, `GoalsStore`, `SubscriptionsStore`, `TransactionsHubStore`) to hold `BigDecimal` values in their respective states.
    - Refactored `TransactionsHubExecutor` to use `fold` with `BigDecimal.ZERO` for calculating day/week/month totals.
    - Added `toEditString()` utility to `BigDecimal` for better handling of numeric input in text fields without scientific notation or trailing zeros.

+++ app/src/main/java/dev/esbi/mizan/feature/addtransaction/presentation/widgets/DynamicAccountSelector.kt
@@ -21,6 +21,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import dev.esbi.mizan.domain.model.Account
 import dev.esbi.mizan.ui.theme.colors.MizanTheme
+import java.math.BigDecimal
 import java.text.NumberFormat
 import java.util.Locale
 
@@ -109,7 +110,7 @@ private fun AccountItem(
     }
 }
 
-private fun formatCurrency(amount: Double, currency: String): String {
+private fun formatCurrency(amount: BigDecimal, currency: String): String {
     val format = NumberFormat.getCurrencyInstance(Locale.US)
     try {
         format.currency = java.util.Currency.getInstance(currency)
